Make executed processes inherit all environment variables
Previous implementation replaced the whole environment with the single
DEMOS_FDS variable. This caused various troubles - for example in the
test suite, we could not execute the processes, because they were not
be in the default PATH.

To make this change easier to implement, we change the forking code
from using vfork to plain fork. We needed vfork in the past, but I
hope that the bugs that required use of vfork are long gone.
Unfortunately, we have no tests (yet), so I just hope.