Do not put demos-sched to a sub-cgroup
When demos-sched puts itself to a cgroup it creates, there is no way
to clean up after itself and remove that cgroup (because populated
cgroups cannot be removed). Therefore, we leave demos where it is and
use a dummy child process to test cgroup pemissions. If executing the
demos-sched with the -s flag, the unified group will be removed by
systemd.