Fixed a possible deadlock during initialization
Before this commit, if a process exited during its initialization window, the scheduler would hang waiting for a completion signal from an already exited process.
To fix this, I set a separate process exit callback during initialization and move on to the next process in case the process exits.
I added a test to `0100-api.t` that should check for this issue.