Silence clang warnings about diagnostic pragmas for spdlog
Clang does not know the `-Wsuggest-attribute=const` warning ignored here, and there doesn't seem to be a flag that would work for both clang and gcc. Simplest solution seems to be to suppress the warning about unknown warning. :)

However, GCC doesn't know -Wunknown-warning-option, so it complains. So we'll also suppress that and now it seems to work correctly and silently on both GCC and Clang.