! Added CPU frequency scaling code in PowerManager
I'm not sure this is ready for PR now - I tested it on i.MX8, but not much on Intel CPUs, which have the added complexity of a custom `intel_pstate` driver with different behavior from the generic governors. I'd appreciate more testing before PR.

Also, there is the question of correctly solving the governor cleanup on crash, which should probably be solved together with cgroup cleanup and other related resources in a unified way (probably using systemd or a separate launcher).