nix: Drop tests with gcc7
Newly added std::filestystem dependency is not available in gcc
version 7 (and earlier). Drop it from tests and keep only test with
gcc8 and later.

This requires a bit of hackery around tests with meson-0.49.2. For
this old meson version, nixpkgs hardcode the compiler to be the
default one from nixpkgs and does not support changing it via stdenv.
The default compiler in that old nixpkgs version happens to be gcc7.
The applied workaround replaces meson command with a shell function,
which unsets the variables with hardcoded compiler so that meson uses
the default gcc and g++ from stdenv.