Implemented issue #41 and changed window and partition ownership
Issue #41 - now, passed message (`-m` param) is printed at the beginning of each window.

MajorFrame is now constructed inside DemosScheduler and both `windows` and `partitions` are moved inside MajorFrame and PartitionManager (respectively). This is part of an ongoing effort to move stuff out of the oversized `main()` where it makes sense.