nix: Include submodules in Nix build sources
Note that this is a workaround until submodules support in
experimental Nix comes to the stable version.

In this commit, all submodules are unconditionally included in the
sources. In next commit, we will make this optional to test builds
both with and without submodules. The reason we don't always want to
use submodules is that compiling submodules significantly increases
compilation times.