Round default budgets to the whole number of milliseconds
Without this, the command added as a test fails with

     Exception: bad conversion

The reason is that the YAML reading code expects budgets to be
integers, but the default budget was computed as float. One of the
reasons for using floats was to use NaN as a special value. When this
float reaches Config::create_demos_objects(), the exception is thrown,
because the budget is expected to be an integer.

Alternatively, we could just use float budgets, but this would make
dumped config files (-d) look ugly ;-)