#57 Switch per-process frequency checks from ahead-of-time in Config to runtime checks in PowerPolicy
The checks here are more benevolent, only reporting an actual collision rather than all potential ones. However, if there's a race condition, the runtime checks may not always catch it.

TODO: if a collision is detected, it throws a runtime exception, which stops the scheduler without a cleanup, leaving behind processes and cgroups