Improve error handling in power policy initialization
When running demos with per-process power policy on NVIDIA TX2, it
fails with the following (not very useful) error:

    demos-sched -p imx8_per_process -C ...
    >>> 11:47:25.545 [info] Activating power management (power policy: 'imx8_per_process', args: '')
    >>> 11:47:25.569 [error] Exception: Unknown power policy selected: imx8_per_process; available policies: high, imx8_alternating, imx8_fixed, imx8_per_process, imx8_per_slice, low, minbe

The reason is that when the power policy reads
.../cpufreq/policy1/scaling_available_frequencies, the read fails with
ENODEV and this somehow translates to the same exception thrown when
we try to select non-existent power policy.

This commit splits error checking into two places. First, we check for
existence of the power policy and then for errors during
initialization of the power policy. Now, when running on TX2, we get
the following output:

    >>> 07:56:49.679 [info] Activating power management (power policy: 'imx8_per_process', args: '')
    >>> 07:56:49.709 [error] Exception: Failed to initialize power policy imx8_per_process: map::at

The "map::at" is still confusing, but definitely better than what we
had before.