cpufreq_policy: Read min/max frequency from cpuinfo files
Previously used scaling_* files do not report frequencies that the
hardware is capable of running at, but a software-based limit for
various governors.

On NVIDIA TX2, those values differ. Specifically,
scaling_max_frequency is set to about 1.4 GHz and cpuinfo_max_freq to
about 2 GHz. In order to use be allowed to use all available
frequencies in DEmOS, we need to use the cpuinfo one. We also need to
update the scaling_max_frequency, which will be done in the next
commit.