#68 Rough support for cyclic execution of BE processes in a window
I don't like that this uses the internal Partition::clear_completed_flag method from Slice. I think that a more architecturally sound solution would be to create BePartition and ScPartition subclasses and move as much of the partition-related logic there, as the special cases for SC and BE are starting to pile up a bit inside Slice.